Created responsive mobile, tablet, and desktop Navbar
Fixed
- Mobile Navbar responsiveness
- Reworked the entire structure, to make it (so I thought at the time) easier to create a mobile navbar. But I shot myself in the foot and continued with the injury
- Clicking elsewhere closes the mobile navbar

Added
- Mobile responsiveness
- An slide-in and out animation for the mobile view
- A new JSX structure that will *eventually* be fixed. It will be fixed when an apple actually falls on Sir Isaac Newton's head
- A new emotion: I'm tired of writing these commit messages. This advancement has been achieved by working forever on a seemingly simple objective. Even Newton theorized for every minute wasted on a damn navbar there is an equal amount of tiredness and I'm gonna do the opposite of what my body's telling me.

Removed
- The old JSX structure
- Energy from my body

To whoever reads this, just now, I'm not okay. This seemingly simple objective, has deprived me of sleep, energy, and time especially on this commit mesage. If this site will be used, no idea. When it will be used, no idea. I just know I've wasted a good solid 10 minutes on this commit message that nobody will read. Ok, I'm done now. I'm going insane. But I honestly want to keep typing, but not sure if it's a good idea because atht's now 11 minutes that I've wasted